requirements.txt
pytest==8.2.2
pytest-json-report==1.5.0
pyyaml==6.0.2
tomli==2.0.1
